src/requirements.txt
python-dotenv
requests
beautifulsoup4
openai
langchain-openai